Change cow circle radii
This changes cow circle radii from being proportional to the number of
cows in a square to being proportional to the square root of the
number of cows in a square.  Since the circle area is proportional to
the square of the radius, this makes the area of circles proportional
to the number of cows, instead of the square of the number of cows.